Switch to 32 bits for simple stack machine

+++ StackMachines.Tests/VM0Tests.fs
@@ -12,22 +12,22 @@ open Swensen.Unquote
 // This file is work in progress.
 
 let setupNoIO code stackContents freeStack : Machine =
-    let data = Seq.concat [code; [EXIT] :> seq<int>; stackContents; {1..freeStack}] |> Seq.map uint64 |> Seq.toArray
+    let data = Seq.concat [code; [EXIT] :> seq<int>; stackContents; {1..freeStack}] |> Seq.map uint32 |> Seq.toArray
     let m = Machine(data,
                     seq { for i in 0 .. 10 do raise (Exception "Unexpected input")},
                     fun _ -> raise (Exception "Unexpected output"))
-    m.StackPointer <- uint64 (Array.length data - freeStack)
+    m.StackPointer <- uint32 (Array.length data - freeStack)
     m
 
 let run (program: int list) (input: string) : string =
     let mutable output: string = ""
 
-    let out (s: seq<uint64>) =
+    let out (s: seq<uint32>) =
         output <- output + System.Text.UTF8Encoding().GetString(Seq.toArray (Seq.map uint8 s))
 
     Machine(
-        program |> Seq.map uint64,
-        Seq.map uint64 input,
+        program |> Seq.map uint32,
+        Seq.map uint32 input,
         out).Run()
 
     output
@@ -40,5 +40,5 @@ let tests =
         testProperty "Add constant" <| fun m n ->
             let machine = setupNoIO (addC m) [n] 1
             machine.Run()
-            test <@ machine.Get (machine.StackPointer - 1UL) = uint64 (m + n) @>
+            test <@ machine.Get (machine.StackPointer - 1u) = uint32 (m + n) @>
     ]
